Added solution for sum of fibonacci numbers
Completed part 3 of the midterm with the soplution sumFibNum.

The solution also implements a debug flag macro to turn on and off
verbose output text

To do next is to figure out what needs to happen in order to use an
execution flag fed through the command line to govern display of this
info or not.