Skeleton for elapsedTime
Program compiles and runs, ans successfully takes input and output and
passes sctructs between the functions.

The algorithm needs a little work. This algorithm should take into
account transitions between 12:00:00 am midnight as the start time and
also deal with what happens if the seconds wrap around 60, the minutes
wrap around, and the hours. Also it should be certain to double check
the order in which times are subtracted.